Markov Process textgenerator with FitnessFunctions
A Short Story Writer was created to produce creative pieces of text based on the architecture of Markov Chains/Processes and a corpus/language provided to the program. Markov Processes are notably reliant on probability, so to avoid generating the same output reoccurring, a ‘Fitness Function’ was implemented, having ‘weightings’ of the functions, similar to that of a value network within neural networks. This allowed the system to increase or decrease certain weight attributes such as ‘Rhyming’ or ‘Punctuation’ in order to bias the text generation to create more creative and interesting texts. The corpus used for A Short Story Writer was the summation of Barack Obamas speeches, as these provided the program with clear, error free text which it uses to generate a new story. The system was also tested on a corpus of the Aladdin story in order to generate a text resembling a story.